+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/scenarios/AccessNonObjectScenario.m
@@ -32,7 +32,7 @@
 @implementation AccessNonObjectScenario
 
 - (void)startBugsnag {
-    self.config.shouldAutoCaptureSessions = NO;
+    self.config.autoTrackSessions = NO;
     [super startBugsnag];
 }